+++ auto-pos-creator/src/external_sc_interactions/pair_actions.rs
@@ -24,41 +24,6 @@ pub struct PairRemoveLiqResult<M: ManagedTypeApi> {
 pub trait PairActionsModule:
     crate::configs::pairs_config::PairsConfigModule + utils::UtilsModule
 {
-    fn buy_half_each_token(
-        &self,
-        input_tokens: EsdtTokenPayment,
-        dest_pair: &ManagedAddress,
-    ) -> DoubleSwapResult<Self::Api> {
-        require!(input_tokens.token_nonce == 0, "Only fungible ESDT accepted");
-        self.require_sc_address(dest_pair);
-
-        let dest_pair_config = self.get_pair_config(dest_pair);
-        let tokens_to_pair_mapper = self.pair_address_for_tokens(
-            &dest_pair_config.first_token_id,
-            &dest_pair_config.second_token_id,
-        );
-        require!(!tokens_to_pair_mapper.is_empty(), "Unknown pair SC");
-
-        let first_amount = &input_tokens.amount / 2u32;
-        let second_amount = &input_tokens.amount - &first_amount;
-
-        let first_swap_tokens = self.perform_tokens_swap(
-            input_tokens.token_identifier.clone(),
-            first_amount,
-            dest_pair_config.first_token_id,
-        );
-        let second_swap_tokens = self.perform_tokens_swap(
-            input_tokens.token_identifier,
-            second_amount,
-            dest_pair_config.second_token_id,
-        );
-
-        DoubleSwapResult {
-            first_swap_tokens,
-            second_swap_tokens,
-        }
-    }
-
     fn perform_tokens_swap(
         &self,
         from_tokens: TokenIdentifier,
